# api/swagger.yaml
definitions:
...
  app.SpaceDTO:
    properties:
      created_at:
        type: integer
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      hardware:
        type: string
      id:
        type: string
      labels:
        $ref: '#/definitions/app.SpaceLabelsDTO'
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      sdk:
        type: string
      updated_at:
        type: integer
      visibility:
        type: string
    type: object
  app.SpaceLabelsDTO:
    properties:
      frameworks:
        items:
          type: string
        type: array
      license:
        type: string
      others:
        items:
          type: string
        type: array
      task:
        type: string
    type: object
  app.SpacesDTO:
    properties:
      spaces:
        items:
          $ref: '#/definitions/repository.SpaceSummary'
        type: array
      total:
        type: integer
    type: object
...
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  controller.reqToCreateSpace:
    properties:
      desc:
        type: string
      fullname:
        type: string
      hardware:
        type: string
      init_readme:
        type: boolean
      license:
        type: string
      name:
        type: string
      owner:
        type: string
      sdk:
        type: string
      visibility:
        type: string
    type: object
...
  controller.reqToUpdateSpace:
    properties:
      desc:
        type: string
      fullname:
        type: string
      name:
        type: string
      visibility:
        type: string
    type: object
...
  controller.spaceDetail:
    properties:
      avatar_id:
        type: string
      created_at:
        type: integer
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      hardware:
        type: string
      id:
        type: string
      labels:
        $ref: '#/definitions/app.SpaceLabelsDTO'
      like_count:
        type: integer
      liked:
        type: boolean
      name:
        type: string
      owner:
        type: string
      sdk:
        type: string
      updated_at:
        type: integer
      visibility:
        type: string
    type: object
  controller.spaceInfo:
    properties:
      avatar_id:
        type: string
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      task:
        type: string
      updated_at:
        type: integer
    type: object
  controller.spacesInfo:
    properties:
      spaces:
        items:
          $ref: '#/definitions/controller.spaceInfo'
        type: array
      total:
        type: integer
    type: object
...
  controller.userSpacesInfo:
    properties:
      avatar_id:
        type: string
      owner:
        type: string
      spaces:
        items:
          $ref: '#/definitions/repository.SpaceSummary'
        type: array
      total:
        type: integer
    type: object
...
  repository.SpaceSummary:
    properties:
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      task:
        type: string
      updated_at:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/space:
    get:
      consumes:
      - application/json
      description: list global public space
      parameters:
      - description: name of space
        in: query
        name: name
        type: string
      - description: task label
        in: query
        name: task
        type: string
      - description: other labels, separate multiple each ones with commas
        in: query
        name: others
        type: string
      - description: license label
        in: query
        name: license
        type: string
      - description: framework labels, separate multiple each ones with commas
        in: query
        name: frameworks
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.spacesInfo'
      summary: ListGlobal
      tags:
      - SpaceWeb
    post:
      consumes:
      - application/json
      description: create space
      parameters:
      - description: body of creating space
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.reqToCreateSpace'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Create
      tags:
      - Space
  /v1/space/{id}:
    delete:
      consumes:
      - application/json
      description: delete space
      parameters:
      - description: id of space
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete
      tags:
      - Space
    put:
      consumes:
      - application/json
      description: update space
      parameters:
      - description: id of space
        in: path
        name: id
        required: true
        type: string
      - description: body of updating space
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.reqToUpdateSpace'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Update
      tags:
      - Space
  /v1/space/{owner}:
    get:
      consumes:
      - application/json
      description: list space
      parameters:
      - description: owner of space
        in: path
        name: owner
        required: true
        type: string
      - description: name of space
        in: query
        name: name
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.userSpacesInfo'
      summary: List
      tags:
      - SpaceWeb
  /v1/space/{owner}/{name}:
    get:
      consumes:
      - application/json
      description: get space
      parameters:
      - description: owner of space
        in: path
        name: owner
        required: true
        type: string
      - description: name of space
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.spaceDetail'
      summary: Get
      tags:
      - SpaceWeb
...
  /v2/space:
    get:
      consumes:
      - application/json
      description: list global public space
      parameters:
      - description: name of space
        in: query
        name: name
        type: string
      - description: task label
        in: query
        name: task
        type: string
      - description: owner of space
        in: query
        name: owner
        required: true
        type: string
      - description: other labels, separate multiple each ones with commas
        in: query
        name: others
        type: string
      - description: license label
        in: query
        name: license
        type: string
      - description: framework labels, separate multiple each ones with commas
        in: query
        name: frameworks
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.SpacesDTO'
      summary: List
      tags:
      - SpaceRestful
  /v2/space/{owner}/{name}:
    get:
      consumes:
      - application/json
      description: get space
      parameters:
      - description: owner of space
        in: path
        name: owner
        required: true
        type: string
      - description: name of space
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.SpaceDTO'
      summary: Get
      tags:
      - SpaceRestful
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and api Bearer.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"